backend/requirements.txt
langgraph==0.4.8
langchain==0.3.26
langchain-google-genai==2.1.5
python-dotenv==1.1.1
langgraph-sdk==0.1.70
langgraph-cli==0.3.3
langgraph-api==0.2.63
fastapi==0.115.13
google-genai==1.21.1
uvicorn==0.34.3
llama-index-embeddings-huggingface==0.5.4
llama-index-vector-stores-postgres==0.5.3
llama-index-readers-file==0.4.9
pymupdf==1.26.3
openai==1.91.0
httpx==0.28.1
structlog==25.4.0
ollama==0.5.1